requirements.txt
streamlit
Pillow
PyMuPDF
altair

